requirements.txt
streamlit==1.23.1
requests==2.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
nltk==3.8.1
python-dotenv==1.0.0
pymongo==4.5.0